+++ circuits/examples/identity/property_check.rs
@@ -161,8 +161,11 @@ impl Relation for CredentialProperty {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: true,
-            sha512: false,
+            sha2_256: true,
+            sha2_512: false,
+            sha3_256: false,
+            keccak_256: false,
+            blake2b: false,
             secp256k1: true,
             bls12_381: false,
             base64: true,