+++ circuits/examples/native_gadget.rs
@@ -92,8 +92,11 @@ impl Relation for NativeGadgetExample {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: false,
-            sha512: false,
+            sha2_256: false,
+            sha2_512: false,
+            sha3_256: false,
+            keccak_256: false,
+            blake2b: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,