+++ circuits/examples/sha_preimage.rs
@@ -45,7 +45,7 @@ impl Relation for ShaPreImageCircuit {
     ) -> Result<(), Error> {
         let witness_bytes = witness.transpose_array();
         let assigned_input = std_lib.assign_many(layouter, &witness_bytes)?;
-        let output = std_lib.sha256(layouter, &assigned_input)?;
+        let output = std_lib.sha2_256(layouter, &assigned_input)?;
         output.iter().try_for_each(|b| std_lib.constrain_as_public_input(layouter, b))
     }
 
@@ -53,8 +53,11 @@ impl Relation for ShaPreImageCircuit {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: true,
-            sha512: false,
+            sha2_256: true,
+            sha2_512: false,
+            sha3_256: false,
+            keccak_256: false,
+            blake2b: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,