+++ circuits/src/lib.rs
@@ -26,6 +26,7 @@ mod utils;
 
 pub mod biguint;
 pub mod ecc;
+pub mod external;
 pub mod field;
 pub mod hash;
 pub mod map;