+++ zkir/src/zkir.rs
@@ -124,8 +124,11 @@ impl Relation for ZkirRelation {
         ZkStdLibArch {
             jubjub: involves_types(&[IrType::JubjubPoint, IrType::JubjubScalar]),
             poseidon: operations.iter().any(|op| matches!(op, Poseidon)),
-            sha256: operations.iter().any(|op| matches!(op, Sha256)),
-            sha512: operations.iter().any(|op| matches!(op, Sha512)),
+            sha2_256: operations.iter().any(|op| matches!(op, Sha256)),
+            sha2_512: operations.iter().any(|op| matches!(op, Sha512)),
+            sha3_256: false,
+            keccak_256: false,
+            blake2b: false,
             secp256k1: false,
             bls12_381: false,
             nr_pow2range_cols: 4,