+++ circuits/CHANGELOG.md
@@ -16,6 +16,7 @@ verification keys break backwards compatibility.
 * Introduce `is_not_equal` and `is_not_equal_to_fixed` [#130](https://github.com/midnightntwrk/midnight-zk/pull/130)
 * Add `to_bytes` and `from_bytes` for `BigUint` [#136](https://github.com/midnightntwrk/midnight-zk/pull/136)
 * Blind limbs of quotient polynomial and ensure ZK [#161](https://github.com/midnightntwrk/midnight-zk/pull/161)
+* Adds support for external crates Keccak and Blake2b [#135](https://github.com/midnightntwrk/midnight-zk/pull/135)
 
 ### Changed
 * Optimize `bigint_to_fe` [#115](https://github.com/midnightntwrk/midnight-zk/pull/115)